backend/requirements.txt
# Django
django==4.0.2
mysqlclient==2.0.3
djangorestframework==3.12.4
django-cors-headers==3.10.1
django-simpleui==2022.1
pytz==2021.3

# Storage
redis==4.0.2

# Celery
# celery==5.2.2

# Server
uwsgi==2.0.20
